Fix decimal json and run a custom eth propose

+++ rewards/classes/RewardsManager.py
@@ -141,17 +141,17 @@ class RewardsManager:
         sett_multipliers = {}
         for sett, user_apy_boosts in self.apy_boosts.items():
             sett_multipliers[sett] = {
-                "min": Decimal(min(user_apy_boosts.values())),
-                "max": Decimal(max(user_apy_boosts.values())),
+                "min": float(min(user_apy_boosts.values())),
+                "max": float(max(user_apy_boosts.values())),
             }
         return sett_multipliers
 
     def get_user_multipliers(self) -> Dict[str, Dict[str, Decimal]]:
         user_multipliers = {}
         for sett, multipliers in self.get_sett_multipliers().items():
-            min_mult = Decimal(multipliers["min"])
-            max_mult = Decimal(multipliers["max"])
-            diff = max_mult - min_mult
+            min_mult = multipliers["min"]
+            max_mult = multipliers["max"]
+            diff = float(max_mult - min_mult)
             for user, boost_info in self.boosts.items():
                 if user not in user_multipliers:
                     user_multipliers[user] = {}
@@ -159,8 +159,8 @@ class RewardsManager:
                 if boost == 1:
                     user_sett_multiplier = min_mult
                 else:
-                    user_sett_multiplier = min_mult + (boost / 2000) * diff
-                user_multipliers[user][sett] = user_sett_multiplier
+                    user_sett_multiplier = min_mult + (float(boost) / 2000) * diff
+                user_multipliers[user][sett] = float(user_sett_multiplier)
 
         return user_multipliers
